requirements.txt
flwr==1.3.0
matplotlib==3.3.3
mysql_connector_repackaged==0.3.1
numpy==1.19.4
opacus==1.3.0
pandas==1.1.4
PyYAML==6.0.1
scikit_learn==0.23.2
setuptools==50.3.2
SQLAlchemy==2.0.19
torch==2.0.1
~andas==0.25.3
alembic>=1.7

